+++ src/components/GoogleSignIn/GoogleSignIn.tsx
@@ -1,6 +1,8 @@
-import React from "react";
 import { signInWithPopup, User } from "firebase/auth";
+import React from "react";
+import googleImage from "../../assets/google.svg";
 import { auth, googleProvider } from "../../services/firebase";
+import Button from "../Button";
 
 interface GoogleSignInProps {
   onUserLogin: (user: User) => void;
@@ -17,12 +19,11 @@ const GoogleSignIn: React.FC<GoogleSignInProps> = ({ onUserLogin }) => {
   };
 
   return (
-    <button
+    <Button
+      label="Continue with Google"
       onClick={signInWithGoogle}
-      className="bg-blue-500 text-white px-4 py-2 rounded-md"
-    >
-      Sign in with Google
-    </button>
+      icon={<img className="h-6 w-6 mr-2" src={googleImage} />}
+    />
   );
 };