+++ src/components/Modal/Modal.tsx
@@ -4,7 +4,8 @@ import {
   DialogPanel,
   DialogTitle,
 } from "@headlessui/react";
-import React, { useRef } from "react";
+import React from "react";
+import Button from "../Button";
 
 interface ModalProps {
   isOpen: boolean;
@@ -23,41 +24,25 @@ const Modal: React.FC<ModalProps> = ({
   onConfirm,
   onCancel,
 }) => {
-  const cancelButtonRef = useRef<HTMLButtonElement>(null);
-
   return (
-    <Dialog
-      open={isOpen}
-      onClose={onCancel}
-      className="relative z-50"
-      initialFocus={cancelButtonRef} // Set initial focus
-    >
+    <Dialog open={isOpen} onClose={onCancel} className="relative z-50">
       <DialogBackdrop className="fixed inset-0 bg-black/30" />
 
-      <div className="fixed inset-0 flex w-screen items-center justify-center rad p-4">
-        <DialogPanel className="w-1/2 max-w-lg space-y-4 border rounded bg-white p-4">
-          <DialogTitle>{title}</DialogTitle>
+      <div className="fixed inset-0 flex w-screen items-center justify-center p-4">
+        <DialogPanel className="w-full max-w-lg space-y-4 border rounded bg-white p-4 shadow-lg dark:bg-gray-800 dark:border-gray-700">
+          <DialogTitle className="text-lg font-semibold text-gray-900 dark:text-gray-200">
+            {title}
+          </DialogTitle>
 
-          <div>{children}</div>
+          <div className="text-gray-700 dark:text-gray-300">{children}</div>
 
-          {/* Actions */}
           <div className="flex gap-2 justify-between">
-            <button
-              ref={cancelButtonRef} // Reference to set initial focus
-              onClick={onCancel}
-              className="mt-4 w-full rounded-md bg-slate-800 py-2 px-4 border border-transparent text-center text-sm text-white transition-all shadow-md hover:shadow-lg focus:bg-slate-700 focus:shadow-none active:bg-slate-700 hover:bg-slate-700 active:shadow-none disabled:pointer-events-none disabled:opacity-50 disabled:shadow-none"
-              type="button"
-            >
-              Cancel
-            </button>
-            <button
-              disabled={isConfirmDisabled}
+            <Button label="Cancel" onClick={onCancel} />
+            <Button
+              label="Confirm"
               onClick={onConfirm}
-              className="mt-4 w-full rounded-md bg-slate-800 py-2 px-4 border border-transparent text-center text-sm text-white transition-all shadow-md hover:shadow-lg focus:bg-slate-700 focus:shadow-none active:bg-slate-700 hover:bg-slate-700 active:shadow-none disabled:pointer-events-none disabled:opacity-50 disabled:shadow-none"
-              type="button"
-            >
-              Confirm
-            </button>
+              disabled={isConfirmDisabled}
+            />
           </div>
         </DialogPanel>
       </div>