+++ src/components/Totalizers/Totalizers.tsx
@@ -1,3 +1,5 @@
+import React from "react";
+
 type TotalizersProps = {
   income: number;
   expense: number;
@@ -8,33 +10,39 @@ const Totalizers: React.FC<TotalizersProps> = ({ income, expense }) => {
     {
       label: "Balance",
       value: income - expense,
-      color: "#475569",
+      colorClass: "text-slate-800 dark:text-slate-400",
     },
     {
       label: "Income",
       value: income,
-      color: "#166534",
+      colorClass: "text-green-800 dark:text-green-400",
     },
     {
       label: "Expense",
       value: expense,
-      color: "#991b1b",
+      colorClass: "text-red-800 dark:text-red-400",
     },
   ];
 
-  const formatNumber = (value: number) => {
-    return new Intl.NumberFormat("pt-BR", {
+  const formatNumber = (value: number) =>
+    new Intl.NumberFormat("pt-BR", {
       style: "currency",
       currency: "BRL",
     }).format(value);
-  };
 
   return (
-    <div className="flex gap-8">
-      {data.map((item) => (
-        <div className="flex flex-col" key={item.label}>
-          <span className="text-base text-slate-800">{item.label}</span>
-          <span className="text-2xl font-light" style={{ color: item.color }}>
+    <div className="flex flex-wrap justify-evenly md:justify-normal gap-4 sm:gap-8 md:py-4">
+      {data.map((item, index) => (
+        <div
+          className={`${
+            index === 0 ? "w-full sm:w-auto" : "w-auto"
+          } flex flex-col items-center sm:items-start text-center sm:text-left`}
+          key={item.label}
+        >
+          <span className="text-base font-light text-slate-600 dark:text-slate-300">
+            {item.label}
+          </span>
+          <span className={`text-2xl ${item.colorClass}`}>
             {formatNumber(item.value)}
           </span>
         </div>